requirements.txt
faiss-cpu==1.8.0.post1
joblib==1.5.3
langgraph==0.2.55
numpy==1.26.4
pandas==2.3.3
python-dotenv==1.1.1
scikit-learn==1.5.1
streamlit==1.54.0
